+++ tests/bootstrap.php
@@ -1,0 +1,14 @@
+<?php 
+/**
+ * PHP simple library for managing Json files.
+ * 
+ * @author     Josantonius - hello@josantonius.com
+ * @copyright  Copyright (c) 2016 - 2017
+ * @license    https://opensource.org/licenses/MIT - The MIT License (MIT)
+ * @link       https://github.com/Josantonius/PHP-Json
+ * @since      1.1.4
+ */
+
+$loader = require __DIR__ . '/../src/bootstrap.php';
+
+$loader->add('Josantonius\Json\Test', __DIR__);